+++ Jornada 2k20/GoBarber/GoBarber-mobile/src/routes/index.tsx
@@ -1,22 +1,23 @@
 import React from 'react';
-import { createStackNavigator } from '@react-navigation/stack';
+import { ActivityIndicator, View } from 'react-native';
 
-import SignIn from '../pages/SignIn';
-import SingUp from '../pages/SingUp';
+import AppRoutes from './app.routes';
+import AuthRoutes from './auth.routes';
 
-// Navegação para autenticação.
-const Auth = createStackNavigator();
+import { useAuth } from '../hooks/auth';
 
-const AuthRoutes: React.FC = () => (
-    <Auth.Navigator
-        screenOptions={{
-            headerShown: false,
-            cardStyle: { backgroundColor: '#312e38' },
-        }}
-    >
-        <Auth.Screen name="SingIn" component={SignIn} />
-        <Auth.Screen name="SingUp" component={SingUp} />
-    </Auth.Navigator>
-);
+const Routes: React.FC = () => {
+  const { user, loading } = useAuth();
 
-export default AuthRoutes;
+  if (loading) {
+    return (
+      <View style={{ flex: 1, justifyContent: 'center', alignItems: 'center' }}>
+        <ActivityIndicator size="large" color="#999" />
+      </View>
+    );
+  }
+
+  return user ? <AppRoutes /> : <AuthRoutes />;
+};
+
+export default Routes;
